+++ crates/cli/src/completions.rs
@@ -1,11 +1,11 @@
 use crate::cli_input::build_cli;
 use crate::errors::CliError;
-use clap_generate::generators::{generate, Bash, Elvish, Fish, PowerShell, Zsh};
+use clap_generate::generators::{Bash, Elvish, Fish, PowerShell, Zsh};
 use clap_generate::Generator;
 use std::io;
 
 /// generates shell specific completions for the cli and prints them to stdout
-pub fn generate_completions(shell: &str) -> Result<(), CliError> {
+pub fn generate(shell: &str) -> Result<(), CliError> {
     match shell {
         "bash" => completions_for_shell(Bash),
         "fish" => completions_for_shell(Fish),
@@ -18,5 +18,5 @@ pub fn generate_completions(shell: &str) -> Result<(), CliError> {
 }
 
 fn completions_for_shell(generator: impl Generator) {
-    generate(generator, &mut build_cli(), "grafbase", &mut io::stdout());
+    clap_complete::generate(generator, &mut build_cli(), "grafbase", &mut io::stdout());
 }