+++ crates/cli/src/dev.rs
@@ -3,13 +3,13 @@ use crate::CliError;
 use common::consts::DEFAULT_PORT;
 use local_gateway::dev_server_api::start_dev_server;
 
-/// cli wrapper for [local_gateway::dev_server_api::start_dev_server]
+/// cli wrapper for [`local_gateway::dev_server_api::start_dev_server`]
 ///
 /// # Errors
 ///
-/// returns [CliError::LocalGatewayError] if the the local gateway returns an error
+/// returns [`CliError::LocalGatewayError`] if the the local gateway returns an error
 ///
-/// returns [CliError::DevServerPanic] if the development server panics
+/// returns [`CliError::DevServerPanic`] if the development server panics
 pub fn dev(search: bool, external_port: Option<u16>) -> Result<(), CliError> {
     trace!("attempting to start dev server");
     let start_port = external_port.unwrap_or(DEFAULT_PORT);
@@ -24,14 +24,14 @@ pub fn dev(search: bool, external_port: Option<u16>) -> Result<(), CliError> {
     server_handle
         .join()
         .map_err(|panic_parameter| match panic_parameter.downcast_ref::<&'static str>() {
-            Some(parameter) => CliError::DevServerPanic(parameter.to_string()),
+            Some(&parameter) => CliError::DevServerPanic(parameter.to_string()),
             None => match panic_parameter.downcast_ref::<String>() {
                 Some(parameter) => CliError::DevServerPanic(parameter.clone()),
                 None => CliError::DevServerPanic("unknown error".to_owned()),
             },
         })?
         .map_err(|error| {
-            report::miniflare_error(error);
+            report::spawned_thread_error(&error.to_string());
             CliError::DevServerPanic("miniflare error".to_owned())
         })?;