+++ crates/common/src/lib.rs
@@ -1,6 +1,11 @@
+/*!
+The common crate provides shared functionality for Grafbase developer tools
+*/
+
 #![forbid(unsafe_code)]
 
 pub mod consts;
 pub mod environment;
 pub mod errors;
 pub mod traits;
+pub mod utils;