+++ crates/dev-server/src/servers.rs
@@ -1,8 +1,10 @@
-use crate::errors::DevServerError;
+use crate::{bridge, errors::DevServerError};
 use common::environment::Environment;
+use common::utils::find_available_port_in_range;
 use rust_embed::RustEmbed;
 use std::{
-    fs, io,
+    fs,
+    ops::Range,
     process::{Command, Output, Stdio},
     thread::{self, JoinHandle},
 };
@@ -15,28 +17,63 @@ struct Assets;
 const WORKER_DIR: &str = "worker";
 const WORKER_FOLDER_VERSION_FILE: &str = "version.txt";
 
+const EPHEMERAL_PORT_RANGE: Range<u16> = 49152..65535;
+
 /// starts a development server by unpacking any files needed by the gateway worker
-/// and starting the miniflare cli in user_grafbase_path in [Environment]
-pub fn start(port: u16) -> Result<JoinHandle<Result<Output, io::Error>>, DevServerError> {
+/// and starting the miniflare cli in `user_grafbase_path` in [`Environment`]
+///
+/// # Errors
+///
+/// returns [`DevServerError::ReadVersion`] if the version file for the extracted worker files cannot be read
+///
+/// returns [`DevServerError::CreateDir`] if the `WORKER_DIR` cannot be created
+///
+/// returns [`DevServerError::WriteFile`] if a file cannot be written into `WORKER_DIR`
+///
+/// # Panics
+///
+/// The spawned server and miniflare thread can panic if either of the two inner spawned threads panic
+pub fn start(port: u16) -> Result<JoinHandle<Result<Output, DevServerError>>, DevServerError> {
     export_embedded_files()?;
+
+    // the bridge looks for an available port within the ephemeral port range and supplies it to the worker,
+    // making the port choice and availability transprent to the user
+    let bridge_port = find_available_port_in_range(EPHEMERAL_PORT_RANGE).ok_or(DevServerError::AvailablePort)?;
 
     trace!("spawining miniflare");
 
     let environment = Environment::get();
 
-    Ok(thread::spawn(move || {
-        Command::new("npx")
-            .arg("--quiet")
-            .arg("miniflare")
-            .arg("--port")
-            .arg(port.to_string())
-            .arg("-c")
-            .arg("wrangler.toml")
-            .current_dir(&environment.user_grafbase_path)
-            .stdout(Stdio::inherit())
-            .stderr(Stdio::inherit())
-            .output()
-    }))
+    let handle = thread::spawn(move || {
+        let bridge_handle = thread::spawn(move || bridge::start(bridge_port));
+
+        let miniflare_handle = thread::spawn(move || {
+            Command::new("npx")
+                .arg("--quiet")
+                .arg("miniflare")
+                .arg("--port")
+                .arg(port.to_string())
+                .arg("-c")
+                .arg("wrangler.toml")
+                .current_dir(&environment.user_grafbase_path)
+                .stdout(Stdio::inherit())
+                .stderr(Stdio::inherit())
+                .output()
+        });
+
+        // unwrapping to propogate any panics
+        bridge_handle.join().unwrap()?;
+
+        let miniflare_output = miniflare_handle
+            .join()
+            /* unwrapping to propogate any panics */
+            .unwrap()
+            .map_err(DevServerError::MiniflareError)?;
+
+        Ok(miniflare_output)
+    });
+
+    Ok(handle)
 }
 
 fn export_embedded_files() -> Result<(), DevServerError> {