+++ crates/local-gateway/src/dev_server_api.rs
@@ -1,15 +1,22 @@
-use std::{io, process::Output, thread};
-
-use crate::{errors::LocalGatewayError, utils::get_availble_port};
+use crate::errors::LocalGatewayError;
 use common::consts::DEFAULT_PORT;
+use common::utils::find_available_port;
+use dev_server::errors::DevServerError;
+use std::{process::Output, thread};
 
 /// starts the dev server if an available port can be found
+///
+/// # Errors
+///
+/// returns [`LocalGatewayError::AvailablePort`] if no available port can  be found
+///
+/// returns [`LocalGatewayError::PortInUse`] if search is off and the supplied port is in use
 pub fn start_dev_server(
     external_port: Option<u16>,
     search: bool,
-) -> Result<(u16, thread::JoinHandle<Result<Output, io::Error>>), LocalGatewayError> {
+) -> Result<(u16, thread::JoinHandle<Result<Output, DevServerError>>), LocalGatewayError> {
     let start_port = external_port.unwrap_or(DEFAULT_PORT);
-    match get_availble_port(search, start_port) {
+    match find_available_port(search, start_port) {
         Some(port) => match dev_server::start(port) {
             Ok(handle) => Ok((port, handle)),
             Err(error) => Err(LocalGatewayError::DevServerError(error)),